+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # CHANGE LOG
 
+## [Unreleased]
+
+### Changed
+
+- Rename DBpedia `redirection_dump` dataset to `entity_redirections`.
+
 ## [2.3.3](https://github.com/binh-vu/kgdata/tree/2.3.3) (2022-07-06)
 
 [Full Changelog](https://github.com/binh-vu/kgdata/compare/1.7.1...2.3.3)
 
-**Changed:**
+### Changed
 
 - Upgrade hugedict from version 1 to version 2 using Rust. Existing RocksDB located in: `<db_folder>/primary` should change to just `<db_folder>`